+++ Kernel/System/HTMLUtils.pm
@@ -14,7 +14,6 @@ use warnings;
 use utf8;
 
 use MIME::Base64;
-use HTML::Truncate;
 
 our @ObjectDependencies = (
     'Kernel::Config',
@@ -1212,99 +1211,6 @@ sub EmbeddedImagesExtract {
     }egxi;
 
     return 1;
-}
-
-=item HTMLTruncate()
-
-truncate an HTML string to certain amount of characters without loosing the HTML tags, the resulting
-string will contain the specified amount of text characters plus the HTML tags, and ellipsis string.
-
-special characters like &aacute; in HTML code are considered as just one character.
-
-    my $HTML = $HTMLUtilsObject->HTMLTruncate(
-        String   => $String,
-        Chars    => 123,
-        Ellipsis => '...',              # optional (defaults to HTML &#8230;) string to indicate
-                                        #    that the HTML was truncated until that point
-        UTF8Mode => 0,                  # optional 1 or 0 (defaults to 0)
-        OnSpace  => 0,                  # optional 1 or 0 (defaults to 0) if enabled, prevents to
-                                        #    truncate in a middle of a word, but in the space before
-    );
-
-returns
-
-    $HTML => 'some HTML code'           # or false in case of a failure
-
-=cut
-
-sub HTMLTruncate {
-    my ( $Self, %Param ) = @_;
-
-    # check needed
-    for my $Needed (qw(String Chars)) {
-
-        if ( !$Param{$Needed} ) {
-            $Kernel::OM->Get('Kernel::System::Log')->Log(
-                Priority => 'error',
-                Message  => "Need $Needed!",
-            );
-            return;
-        }
-    }
-
-    # translate params for compatibility reasons with HTML::Truncate
-    my %CompatibilityParams = (
-        'utf8_mode' => $Param{UTF8Mode} ? 1 : 0,
-        'on_space'  => $Param{OnSpace}  ? 1 : 0,
-        'chars'     => $Param{Chars},
-        'repair'    => 1,
-    );
-
-    if ( defined $Param{Ellipsis} ) {
-        $CompatibilityParams{ellipsis} = $Param{Ellipsis};
-    }
-
-    # create new HTML truncate object (with the specified options)
-    my $HTMLTruncateObject;
-    eval {
-        $HTMLTruncateObject = HTML::Truncate->new(%CompatibilityParams);
-    };
-
-    if ( !$HTMLTruncateObject ) {
-        $Kernel::OM->Get('Kernel::System::Log')->Log(
-            Priority => 'error',
-            Message  => "Could not create HTMLTruncateObject: $@",
-        );
-        return;
-    }
-
-    # sanitize the string
-    my %Safe = $Self->Safety(
-        String         => $Param{String},
-        NoApplet       => 1,
-        NoObject       => 1,
-        NoEmbed        => 1,
-        NoSVG          => 1,
-        NoImg          => 1,
-        NoIntSrcLoad   => 1,
-        NoExtSrcLoad   => 1,
-        NoJavaScript   => 1,
-        ReplacementStr => '✂︎',
-    );
-
-    # truncate the HTML input string
-    my $Result;
-    if ( !eval { $Result = $HTMLTruncateObject->truncate( $Safe{String} ) } ) {
-
-        $Kernel::OM->Get('Kernel::System::Log')->Log(
-            Priority => 'error',
-            Message  => 'Truncating string failed: ' . $@,
-        );
-
-        return;
-    }
-
-    return $Result;
 }
 
 1;