+++ Kernel/System/TemplateGenerator.pm
@@ -756,88 +756,6 @@ sub NotificationEvent {
         last ARTICLE;
     }
 
-    # get ID for customer and agent articles
-    my $CustomerArticleID = $Article{ArticleID}      || '';
-    my $AgentArticleID    = $ArticleAgent{ArticleID} || '';
-
-    # flag to see if an HTMLBody for Customer is present
-    my $CustomerHTMLBodyPresent = 0;
-
-    # get articles for later use
-    my @ArticleBox = $TicketObject->ArticleContentIndex(
-        TicketID                   => $Param{TicketID},
-        DynamicFields              => 0,
-        UserID                     => $Param{UserID},
-        StripPlainBodyAsAttachment => 2,
-    );
-
-    if ( $Self->{RichText} ) {
-        ARTICLEBOX:
-        for my $ArticleItem ( reverse @ArticleBox ) {
-
-            if ( $CustomerArticleID ne $ArticleItem->{ArticleID} && $AgentArticleID ne $ArticleItem->{ArticleID} ) {
-                next ARTICLEBOX;
-            }
-
-            if ( $ArticleItem->{AttachmentIDOfHTMLBody} ) {
-
-                # get a attachment
-                my %Data = $TicketObject->ArticleAttachment(
-                    ArticleID => $ArticleItem->{ArticleID},
-                    FileID    => $ArticleItem->{AttachmentIDOfHTMLBody},
-                    UserID    => $Param{UserID},
-                );
-
-                # get charset and convert content to internal charset
-                my $Charset;
-                if ( $Data{ContentType} =~ m/.+?charset=("|'|)(.+)/ig ) {
-                    $Charset = $2;
-                    $Charset =~ s/"|'//g;
-                }
-                if ( !$Charset ) {
-                    $Charset = 'us-ascii';
-                    $Data{ContentType} .= '; charset="us-ascii"';
-                }
-
-                # convert charset
-                if ($Charset) {
-                    $Data{Content} = $Kernel::OM->Get('Kernel::System::Encode')->Convert(
-                        Text => $Data{Content},
-                        From => $Charset,
-                        To   => 'utf-8',
-                    );
-
-                    # replace charset in content
-                    $Data{ContentType} =~ s/\Q$Charset\E/utf-8/gi;
-                    $Data{Content} =~ s/(charset=("|'|))\Q$Charset\E/$1utf-8/gi;
-                }
-
-                $Data{Content} =~ s/&amp;/&/g;
-                $Data{Content} =~ s/&lt;/</g;
-                $Data{Content} =~ s/&gt;/>/g;
-                $Data{Content} =~ s/&quot;/"/g;
-
-                # strip head, body and meta elements
-                my $HTMLBody = $Kernel::OM->Get('Kernel::System::HTMLUtils')->DocumentStrip(
-                    String => $Data{Content},
-                );
-
-                # set HTML body for customer article
-                if ( $CustomerArticleID eq $ArticleItem->{ArticleID} ) {
-                    $Param{CustomerMessageParams}->{HTMLBody} = $HTMLBody;
-
-                    # set flag for customer HTML body
-                    $CustomerHTMLBodyPresent = 1;
-                }
-
-                # set HTML body for agent article
-                if ( $AgentArticleID eq $ArticleItem->{ArticleID} ) {
-                    $ArticleAgent{HTMLBody} = $HTMLBody;
-                }
-            }
-        }
-    }
-
     # get  HTMLUtils object
     my $HTMLUtilsObject = $Kernel::OM->Get('Kernel::System::HTMLUtils');
 
@@ -880,20 +798,6 @@ sub NotificationEvent {
     # copy the correct language message attributes to a flat structure
     for my $Attribute (qw(Subject Body ContentType)) {
         $Notification{$Attribute} = $Notification{Message}->{$Language}->{$Attribute};
-    }
-
-    # convert values to HTML to get correct line breaks etc.
-    if ( $Notification{ContentType} =~ m{text\/html} && $CustomerHTMLBodyPresent ) {
-        KEY:
-        for my $Key ( sort keys %{ $Param{CustomerMessageParams} || {} } ) {
-            next KEY if !$Param{CustomerMessageParams}->{$Key};
-
-            # convert HTMLBody to HTML is not needed
-            next KEY if $Key eq 'HTMLBody';
-            $Param{CustomerMessageParams}->{$Key} = $HTMLUtilsObject->ToHTML(
-                String => $Param{CustomerMessageParams}->{$Key},
-            );
-        }
     }
 
     for my $Key (qw(From To Cc Subject Body ContentType ArticleType)) {
@@ -1389,7 +1293,6 @@ sub _Replace {
         # html quoting of content
         if (
             $Param{RichText}
-            && !$Data{HTMLBody}
             && ( !$Data{ContentType} || $Data{ContentType} !~ /application\/json/ )
             )
         {
@@ -1403,6 +1306,7 @@ sub _Replace {
                 );
             }
         }
+
         if (%Data) {
 
             # Check if content type is JSON
@@ -1441,47 +1345,32 @@ sub _Replace {
             #   <OTRS_CUSTOMER_BODY[n]>, <OTRS_AGENT_EMAIL[n]>..., <OTRS_COMMENT>
             if ( $Param{Text} =~ /$Start(?:(?:$DataType(EMAIL|NOTE|BODY)\[(.+?)\])|(?:OTRS_COMMENT))$End/g ) {
 
-                my $Line = $2 || 2500;
+                my $Line       = $2 || 2500;
                 my $NewOldBody = '';
-                if ( $Data{HTMLBody} ) {
+                my @Body       = split( /\n/, $Data{Body} );
 
-                    # comment
-                    my $CharactersPerLine = $ConfigObject->Get('Notification::CharactersPerLine') || 80;
-                    my $CharactersLong = $Line * $CharactersPerLine; # 80 is a fixed value for testing, it should change
+                for my $Counter ( 0 .. $Line - 1 ) {
 
-                    $NewOldBody = $Kernel::OM->Get('Kernel::System::HTMLUtils')->HTMLTruncate(
-                        String   => $Data{HTMLBody},
-                        Chars    => $CharactersLong,
-                        Ellipsis => '...',
-                        UTF8Mode => 1,
-                        OnSpace  => 1,
-                    );
-                }
-                else {
-                    my @Body = split( /\n/, $Data{Body} );
-                    for my $Counter ( 0 .. $Line - 1 ) {
+                    # 2002-06-14 patch of Pablo Ruiz Garcia
+                    # http://lists.otrs.org/pipermail/dev/2002-June/000012.html
+                    if ( $#Body >= $Counter ) {
 
-                        # 2002-06-14 patch of Pablo Ruiz Garcia
-                        # http://lists.otrs.org/pipermail/dev/2002-June/000012.html
-                        if ( $#Body >= $Counter ) {
-
-                            # add no quote char, do it later by using DocumentCleanup()
-                            if ( $Param{RichText} ) {
-                                $NewOldBody .= $Body[$Counter];
-                            }
-
-                            # add "> " as quote char
-                            else {
-                                $NewOldBody .= "> $Body[$Counter]";
-                            }
-
-                            # add new line
-                            if ( $Counter < ( $Line - 1 ) ) {
-                                $NewOldBody .= "\n";
-                            }
+                        # add no quote char, do it later by using DocumentCleanup()
+                        if ( $Param{RichText} ) {
+                            $NewOldBody .= $Body[$Counter];
                         }
-                        $Counter++;
+
+                        # add "> " as quote char
+                        else {
+                            $NewOldBody .= "> $Body[$Counter]";
+                        }
+
+                        # add new line
+                        if ( $Counter < ( $Line - 1 ) ) {
+                            $NewOldBody .= "\n";
+                        }
                     }
+                    $Counter++;
                 }
 
                 chomp $NewOldBody;
@@ -1501,6 +1390,7 @@ sub _Replace {
                     );
                 }
 
+                # replace tag
                 $Param{Text}
                     =~ s/$Start(?:(?:$DataType(EMAIL|NOTE|BODY)\[(.+?)\])|(?:OTRS_COMMENT))$End/$NewOldBody/g;
             }